chore: refactor before merge
