+++ Mail/Views/New Message/Signatures/SignaturesManager.swift
@@ -1,0 +1,66 @@
+//
+/*
+ Infomaniak Mail - iOS App
+ Copyright (C) 2022 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import MailCore
+import SwiftUI
+
+/// Something to track the initial loading of a default signature
+//@StateObject private var signatureManager: SignaturesManager
+
+final class SignaturesManager: ObservableObject {
+    @Published var doneLoadingDefaultSignature = false
+
+    private let mailboxManager: MailboxManager
+    init(mailboxManager: MailboxManager) {
+        self.mailboxManager = mailboxManager
+
+        loadSignaturesIfNeeded()
+    }
+
+    private func loadSignaturesIfNeeded() {
+        Task {
+            do {
+                // check has default
+                guard mailboxManager.getStoredSignatures().default == nil else {
+                    await MainActor.run {
+                        doneLoadingDefaultSignature = true
+                    }
+                    return
+                }
+
+                // load list
+                try await mailboxManager.refreshAllSignatures()
+                assert(mailboxManager.getStoredSignatures().default != nil, "Expecting a default signature")
+
+                await MainActor.run {
+                    doneLoadingDefaultSignature = true
+                }
+
+            } catch {
+                await MainActor.run {
+                    // We want to make sure that in the end the user will not be blocked.
+                    doneLoadingDefaultSignature = true
+                }
+
+                print("error : \(error)")
+                // TODO: sentry
+            }
+        }
+    }
+}
